Enabled part of GUIStates for xbox and non tool loads.  Mainly just going to use the state pool for non tool
SaveToFile using filestream instead of title file

Fix non tool matlib load to use shared content

PurgeStates routine, though dunno if I will use it.  The extra states aren't saved.

BSPTestDebug configuration so it can get to the shared assets

Typo in mapvis

SharedContent in BSPTest

Fix readonly columns in material grid

Fix ReadVertexDeclaration to work with 4.0.  Some of the stuff that used to be int16 and byte are now just int.  Thanks for the warning.

+++ UtilityLib/FileUtil.cs
@@ -116,12 +116,12 @@ namespace UtilityLib
 			VertexElement	[]vels	=new VertexElement[numElements];
 			for(int i=0;i < numElements;i++)
 			{
-				short	offset		=br.ReadInt16();
+				int	offset		=br.ReadInt32();
 
 				VertexElementFormat	vef	=(VertexElementFormat)br.ReadUInt32();
 				VertexElementUsage	veu	=(VertexElementUsage)br.ReadUInt32();
 
-				byte	usageIndex	=br.ReadByte();
+				int	usageIndex	=br.ReadInt32();
 
 				vels[i]	=new VertexElement(offset, vef, veu, usageIndex);
 			}
